requirements.txt
# Core web
Flask==3.0.3
gunicorn==23.0.0
flask-cors==4.1.2

# Core utils
attrs==25.4.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
humanfriendly==10.0
packaging==25.0
platformdirs==4.5.1
typing_extensions==4.15.0
python-dotenv==1.2.1

# HTTP
requests==2.32.3
urllib3==2.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11

# Flask deps
Werkzeug==3.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3

# JSON / schema
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# Math / science
numpy==1.26.4
scipy==1.15.3
sympy==1.14.0
mpmath==1.3.0
networkx==3.4.2

# Image / ML (COMPATIBLE)
opencv-python-headless==4.9.0.80
pillow==10.4.0
imageio==2.37.2
scikit-image==0.25.2
tifffile==2025.5.10
lazy_loader==0.4

# AI / rembg
onnxruntime==1.23.2
numba==0.59.1
llvmlite==0.42.0
flatbuffers==25.12.19
protobuf==6.33.2
PyMatting==1.1.14
rembg==2.0.69

# Misc
pooch==1.8.2
tqdm==4.67.1
pyreadline3==3.5.4
